+++ src/webgl/shaderity_shaders/ColorGradingUsingLUTsShader/ColorGradingUsingLUTsShader.vert
@@ -15,9 +15,9 @@
 void main(){
 /* shaderity: @{mainPrerequisites} */
 
-  mat4 worldMatrix = get_worldMatrix(a_instanceInfo.x);
-  mat4 viewMatrix = get_viewMatrix(cameraSID);
-  mat4 projectionMatrix = get_projectionMatrix(cameraSID);
+  mat4 worldMatrix = get_worldMatrix(uint(a_instanceInfo.x));
+  mat4 viewMatrix = get_viewMatrix(uint(cameraSID));
+  mat4 projectionMatrix = get_projectionMatrix(uint(cameraSID));
   gl_Position = projectionMatrix * viewMatrix * worldMatrix * a_position;
 
   v_texcoord_0 = a_texcoord_0;
