+++ src/webgl/shaderity_shaders/DepthMomentEncodeShader/DepthMomentEncodeShader.vert
@@ -24,11 +24,11 @@ void main()
 
 /* shaderity: @{mainPrerequisites} */
 
-  mat4 worldMatrix = get_worldMatrix(a_instanceInfo.x);
-  mat4 viewMatrix = get_viewMatrix(cameraSID);
-  mat4 projectionMatrix = get_projectionMatrix(cameraSID);
-  mat3 normalMatrix = get_normalMatrix(a_instanceInfo.x);
-  bool isBillboard = get_isBillboard(a_instanceInfo.x);
+  mat4 worldMatrix = get_worldMatrix(uint(a_instanceInfo.x));
+  mat4 viewMatrix = get_viewMatrix(uint(cameraSID));
+  mat4 projectionMatrix = get_projectionMatrix(uint(cameraSID));
+  mat3 normalMatrix = get_normalMatrix(uint(a_instanceInfo.x));
+  bool isBillboard = get_isBillboard(uint(a_instanceInfo.x));
 
   v_color = a_color;
 
@@ -50,7 +50,7 @@ void main()
 
   gl_Position = projectionMatrix * viewMatrix * v_position_inWorld;
 
-  bool visibility = get_isVisible(a_instanceInfo.x);
+  bool visibility = get_isVisible(uint(a_instanceInfo.x));
   if (!visibility)
   {
     gl_Position = vec4(0.0);
